+++ sites/geohub/src/components/maplibre/raster/RasterSimpleLegend.svelte
@@ -128,6 +128,7 @@
 	});
 
 	const getPreviewUrl = (width: number, height: number) => {
+		if (!$map.getLayer(layerId)) return;
 		let titilerBaseUrl = links.find((l) => l.rel === 'cog')?.href;
 		if (!titilerBaseUrl) return;
 
@@ -232,29 +233,31 @@
 		{/if}
 	{:else}
 		{@const layerStyle = $map.getStyle().layers.find((l) => l.id === layerId)}
-		{#if layerStyle.type === 'hillshade'}
-			{@const accentColor = getHillshadeColor()}
-			{@const shadowColor = getHillshadeShadowColor()}
-			<div class="is-flex is-align-items-center">
-				<span class="icon is-normal">
-					<i
-						class="fa-solid fa-wave-square fa-xl"
-						style="color: {accentColor}; text-shadow: 2px 2px {shadowColor};"
-					></i>
-				</span>
-				<span class="ml-3 is-size-6">Hillshade</span>
-			</div>
-		{:else}
-			{@const previewUrl = getPreviewUrl(64, 64)}
-			{#key isLayerChanged}
-				{#if previewUrl}
-					<figure class="image is-64x64">
-						<img src={previewUrl} alt={algorithmId} width="64" height="64" />
-					</figure>
-				{:else}
-					<span>No preview is available</span>
-				{/if}
-			{/key}
+		{#if layerStyle}
+			{#if layerStyle.type === 'hillshade'}
+				{@const accentColor = getHillshadeColor()}
+				{@const shadowColor = getHillshadeShadowColor()}
+				<div class="is-flex is-align-items-center">
+					<span class="icon is-normal">
+						<i
+							class="fa-solid fa-wave-square fa-xl"
+							style="color: {accentColor}; text-shadow: 2px 2px {shadowColor};"
+						></i>
+					</span>
+					<span class="ml-3 is-size-6">Hillshade</span>
+				</div>
+			{:else}
+				{@const previewUrl = getPreviewUrl(64, 64)}
+				{#key isLayerChanged}
+					{#if previewUrl}
+						<figure class="image is-64x64">
+							<img src={previewUrl} alt={algorithmId} width="64" height="64" />
+						</figure>
+					{:else}
+						<span>No preview is available</span>
+					{/if}
+				{/key}
+			{/if}
 		{/if}
 	{/if}
 {:else}