+++ sites/geohub/src/components/util/stac/StacCatalogCollections.svelte
@@ -1,15 +1,28 @@
 <script lang="ts">
+	import { ALGORITHM_TAG_KEY } from '$components/maplibre/raster/RasterAlgorithmExplorer.svelte';
 	import StacCollectionMap from '$components/util/stac/StacCollectionMap.svelte';
-	import type { StacCatalogBreadcrumb, StacCollection } from '$lib/types';
+	import type { DatasetFeature, StacCatalogBreadcrumb, StacCollection } from '$lib/types';
+	import { Tabs, type Tab } from '@undp-data/svelte-undp-components';
 	import { createEventDispatcher, onMount } from 'svelte';
+	import StacCatalogTools from './StacCatalogTools.svelte';
 
 	const dispatch = createEventDispatcher();
 
 	export let stacId: string;
 	export let collectionUrl: string;
 	export let url: string;
+	export let dataset: DatasetFeature = undefined;
 
-	let collections: StacCollection;
+	let collection: StacCollection;
+
+	let tabs: Tab[] = [{ id: 'catalog', label: 'Explore from catalog' }];
+
+	const algorithmTags = dataset?.properties.tags?.filter((t) => t.key === ALGORITHM_TAG_KEY) ?? [];
+	if (algorithmTags.length > 0) {
+		tabs.push({ id: 'tools', label: 'Tools' });
+	}
+
+	let activeTab = tabs[0].id;
 
 	onMount(() => {
 		initialise();
@@ -18,8 +31,8 @@
 	$: url, initialise();
 
 	const initialise = async () => {
-		collections = undefined;
-		collections = await fetchCollection(url);
+		collection = undefined;
+		collection = await fetchCollection(url);
 	};
 
 	const fetchCollection = async (collectionUrl: string) => {
@@ -37,15 +50,37 @@
 	};
 </script>
 
-{#if collections}
-	<p class="is-size-6 mb-4">{collections.description}</p>
+{#if collection}
+	<p class="is-size-6 mb-4">{collection.description}</p>
 
-	<StacCollectionMap
-		bind:stacId
-		bind:collectionUrl
-		bind:url
-		bind:links={collections.links}
-		on:selected={handleChildSelected}
-		on:dataAdded={dataAddedToMap}
-	/>
+	{#if algorithmTags.length > 0}
+		<Tabs
+			bind:tabs
+			bind:activeTab
+			isCentered={false}
+			isBoxed={false}
+			isUppercase={true}
+			fontWeight="bold"
+		/>
+	{/if}
+
+	<div hidden={activeTab !== 'catalog'}>
+		<StacCollectionMap
+			bind:stacId
+			bind:collectionUrl
+			bind:url
+			bind:links={collection.links}
+			on:selected={handleChildSelected}
+			on:dataAdded={dataAddedToMap}
+		/>
+	</div>
+
+	<div hidden={activeTab !== 'tools'}>
+		<StacCatalogTools
+			bind:collectionUrl
+			bind:collection
+			bind:dataset
+			on:dataAdded={dataAddedToMap}
+		/>
+	</div>
 {/if}