+++ sites/geohub/src/lib/server/defaultStyle/RasterDefaultStyle.ts
@@ -31,8 +31,8 @@ export default class RasterDefaultStyle implements DefaultStyleTemplate {
 		this.bandIndex = bandIndex;
 	}
 
-	public create = async (colormap_name?: string) => {
-		this.metadata = await this.getMetadata();
+	public create = async (colormap_name?: string, algorithmId?: string) => {
+		this.metadata = await this.getMetadata(algorithmId);
 		if (!this.bandIndex) {
 			this.bandIndex = getActiveBandIndex(this.metadata);
 		}
@@ -112,10 +112,18 @@ export default class RasterDefaultStyle implements DefaultStyleTemplate {
 			}
 
 			titilerApiUrlParams = {
-				bidx: this.bandIndex + 1,
 				rescale: rescale.join(','),
 				colormap_name: colormap
 			};
+
+			const algoId = this.dataset.properties?.tags?.find(
+				(t) => t.key === 'algorithm' && t.value === algorithmId
+			)?.value;
+			if (algoId) {
+				titilerApiUrlParams['algorithm'] = algoId;
+			} else {
+				titilerApiUrlParams['bidx'] = this.bandIndex + 1;
+			}
 
 			Object.keys(titilerApiUrlParams).forEach((key) => {
 				params.set(key, `${titilerApiUrlParams[key]}`);
@@ -187,7 +195,7 @@ export default class RasterDefaultStyle implements DefaultStyleTemplate {
 		return data;
 	};
 
-	public getMetadata = async () => {
+	public getMetadata = async (algorithmId?: string) => {
 		const metadataUrl = this.dataset.properties?.links?.find((l) => l.rel === 'info').href;
 		if (!metadataUrl) return this.metadata;
 		const res = await fetch(metadataUrl);
@@ -199,10 +207,13 @@ export default class RasterDefaultStyle implements DefaultStyleTemplate {
 			const resStatistics = await fetch(
 				`${
 					this.dataset.properties.links.find((l) => l.rel === 'statistics').href
-				}&histogram_bins=10`
+				}&histogram_bins=10${algorithmId ? `&algorithm=${algorithmId}` : ''}`
 			);
+			if (!resStatistics.ok) {
+				error(resStatistics.status, resStatistics.statusText);
+			}
 			const statistics = await resStatistics.json();
-			if (statistics) {
+			if (statistics && !algorithmId) {
 				for (let i = 0; i < this.metadata.band_metadata.length; i++) {
 					const bandValue = this.metadata.band_metadata[i][0] as string;
 					const bandDetails = statistics[bandValue];
@@ -220,6 +231,30 @@ export default class RasterDefaultStyle implements DefaultStyleTemplate {
 						meta['STATISTICS_MEDIAN'] = bandDetails.median;
 					}
 				}
+				this.metadata.stats = statistics;
+			} else if (statistics && algorithmId) {
+				// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+				// @ts-ignore
+				this.metadata.band_metadata = [];
+				this.metadata.band_descriptions = [];
+				Object.keys(statistics).forEach((bName) => {
+					const bandDetails = statistics[bName];
+					const bandMeta: BandMetadata = {
+						STATISTICS_MAXIMUM: bandDetails.max,
+						STATISTICS_MEAN: bandDetails.mean,
+						STATISTICS_MINIMUM: bandDetails.min,
+						STATISTICS_STDDEV: bandDetails.std,
+						STATISTICS_VALID_PERCENT: bandDetails.STATISTICS_VALID_PERCENT,
+						STATISTICS_MEDIAN: bandDetails.median
+					};
+					// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+					// @ts-ignore
+					this.metadata.band_metadata.push([bName, bandMeta]);
+					// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+					// @ts-ignore
+					this.metadata.band_descriptions.push([bName, '']);
+				});
+				this.metadata.stats = statistics;
 			}
 		}